requirements.txt
timm
torchvision
Pillow
SimpleITK
numpy
